♻ refactor: 패키지 리팩터링 (2023.07.11 임정민)
