+++ server/src/main/java/competnion/global/auth/dto/LoginDto.java
@@ -1,4 +1,4 @@
-package competnion.global.security.dto;
+package competnion.global.auth.dto;
 
 import lombok.Getter;
